Fixing issues identified by Courtney
