+++ assets/js/pages/relationship.js
@@ -1,0 +1,4 @@
+
+const checkUnknowns = require('./checkUnknowns');
+
+checkUnknowns('no_working_details', 'stopWorking','2');
