+++ config/locales/en.json
@@ -457,11 +457,15 @@
 	"condition.save_condition": "Save condition",
 	"Save": "Save",
 	"Save and add another condition": "Save and add another condition",
+	"For example: Feb 2008": "For example: Feb 2008",
+	"For example: Feb 3, 2016": "For example: Feb 3, 2016",
+	"For example: Nov 15, 2019": "For example: Nov 15, 2019",
+	"Did not discuss": "Did not discuss",
 	"dashboard.title": "Medical report for the Canada Pension Plan Disability Benefits",
 	"dashboard.patient": "Patient: ",
 	"dashboard.save-and-return": "Save and return later",
 	"dashboard.submit": "Submit report to Service Canada",
-	"dashboard.basic_info": "Basic Patient Information (TBC)",
+	"dashboard.basic_info": "Basic Patient Information",
 	"dashboard.functional-limits": "Functional limitations and restrictions",
 	"dashboard.medical-conditions": "Medical conditions",
 	"dashboard.medications": "Medications",
@@ -469,10 +473,10 @@
 	"dashboard.future-work-capacity": "Future work capacity",
 	"dashboard.supporting-docs": "Supporting documents",
 	"dashboard.declaration": "Your information and declaration",
-  "dashboard.intro1": "Provide as much information about ",
-  "dashboard.intro2": " as you can, even if you do not think it is relevant.",
-  "dashboard.intro3": "Service Canada will decide if",
-  "dashboard.intro4":  "is eligible for the Canada Pension Plan Disability Benefits, based on the detail you provide.",
+	"dashboard.intro1": "Provide as much information about ",
+	"dashboard.intro2": " as you can, even if you do not think it is relevant.",
+	"dashboard.intro3": "Service Canada will decide if",
+	"dashboard.intro4": "is eligible for the Canada Pension Plan Disability Benefits, based on the detail you provide.",
 	"dashboard.pay": "Service Canada will pay up to <b>$85 directly</b> to you once you’ve completed this medical report.",
 	"dashboard.pay_link": "Claim $85 compensation",
 	"dashboard.complete-sections": "Complete the following sections:",
@@ -481,9 +485,5 @@
 	"dashboard.not-completed": "Not Completed",
 	"dashboard.consent": "has given consent for Service Canada to obtain personal information about them.",
 	"dashboard.view_download_consent": "View or download their consent as a PDF for your records.",
-	"return_to_dashboard": "Save",
-	"For example: Feb 2008": "For example: Feb 2008",
-	"For example: Feb 3, 2016": "For example: Feb 3, 2016",
-	"For example: Nov 15, 2019": "For example: Nov 15, 2019",
-	"Did not discuss": "Did not discuss"
+	"return_to_dashboard": "Save"
 }
