+++ config/webpack.js
@@ -44,6 +44,7 @@ module.exports.webpack = {
         app: './assets/js/app.js',
         styles: './assets/styles/app.scss',
         condition: './assets/js/pages/condition.js',
+        relationship: './assets/js/pages/relationship.js',
         conditionAdder: './assets/js/pages/conditionAdder.js'
       },
       output: {
