+++ views/pages/dashboard.njk
@@ -50,7 +50,7 @@
   <div class="lg:w-4/5">
     <h2> {{ __('dashboard.complete-sections') }} </h2>
     <hr class="my-4"/>
-    {{ sectionLink('dashboard','dashboard.basic_info', sectionsCompleted.personal) }}
+    {{ sectionLink('relationship','dashboard.basic_info', sectionsCompleted.personal) }}
     <hr class="my-4"/>
     {{ sectionLink('dashboard','dashboard.functional-limits', sectionsCompleted.functional)}}
     <hr class="my-4"/>
